requirements.txt
keras==2.7.0
numpy==1.20.1
opencv_python==4.6.0.66
Pillow==9.4.0
streamlit==1.5.1
